Changed storage file sizes to longs to allow bigger uploads

+++ Shared/PluralizeHelpers.cs
@@ -2,12 +2,12 @@ namespace ThriveDevCenter.Shared
 {
     public static class PluralizeHelpers
     {
-        public static string PrintCount(this string value, int count)
+        public static string PrintCount(this string value, long count)
         {
             return $"{count} {value.Pluralize(count)}";
         }
 
-        public static string Pluralize(this string value, int count)
+        public static string Pluralize(this string value, long count)
         {
             if (count == 1)
                 return value;
